feat: settings dialog
