+++ src/AppRoutes.tsx
@@ -15,7 +15,6 @@ import { RecordingPreviewer } from '@/views/RecordingPreviewer'
 import { Generator } from '@/views/Generator/Generator'
 import { Validator } from '@/views/Validator'
 import { routeMap } from './routeMap'
-import { Settings } from './views/Settings/Settings'
 import { ErrorElement } from './ErrorElement'
 
 const router = createHashRouter(
@@ -26,7 +25,6 @@ const router = createHashRouter(
       errorElement={<ErrorElement />}
     >
       <Route index element={<Home />} />
-      <Route path={routeMap.settings} element={<Settings />} />
       <Route path={routeMap.recorder} element={<Recorder />} />
       <Route
         path={routeMap.recordingPreviewer}
