+++ src/components/Settings/SettingsSection.tsx
@@ -8,21 +8,20 @@ type SettingsSectionProps = {
 
 export function SettingsSection({ title, children }: SettingsSectionProps) {
   return (
-    <Flex gap="2" direction="column" px="3">
+    <Flex gap="2" direction="column" p="1">
       <Heading
         size="2"
-        mt="4"
+        my="4"
         css={css`
           font-size: 18px;
           line-height: 24px;
           font-weight: 500;
-          padding: var(--space-2);
         `}
       >
         {title}
       </Heading>
 
-      <Box mx="2">{children}</Box>
+      <Box>{children}</Box>
     </Flex>
   )
 }
